// frontend/src/components/RichText/functions/types/Active_format.type.ts
export type ActiveFormatType = {
  bold?: boolean;
  italic?: boolean;
  underline?: boolean;
  heading1?: boolean;
  heading2?: boolean;
  heading3?: boolean;
  heading4?: boolean;
  heading5?: boolean;
  heading6?: boolean;
  bulletList?: boolean;
  numberedList?: boolean;
  quote?: boolean;
  code?: boolean;
};